[S06P31C203-293]Fix: 음성메세지 기능 수정

See merge request s06-final/S06P31C203!38

+++ BE/src/main/java/com/ssafy/blahblah/api/request/member/ReportAnsPostReq.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ReportAnsPostReq {
-    Long day;
+    int day;
     String reason;
 }
